more work

+++ common/src/main/java/net/moist/mixin/SpreadingSnowyDirtBlockMixin.java
@@ -26,6 +26,25 @@ public abstract class SpreadingSnowyDirtBlockMixin extends SnowyDirtBlock {
 	@Shadow private static boolean canBeGrass(BlockState blockState, LevelReader levelReader, BlockPos blockPos) {
 		return true;
 	}
+//	@Unique private static boolean tough_terrain$canBeGrassDebug(BlockState blockState, LevelReader levelReader, BlockPos blockPos) {
+//		BlockPos blockPos2 = blockPos.above();
+//		BlockState blockState2 = levelReader.getBlockState(blockPos2);
+//		if (blockState2.is(Blocks.SNOW) && blockState2.getValue(SnowLayerBlock.LAYERS) == 1) {
+//			System.out.println("true at"+blockPos+" due to snow layers only being one");
+//			return true;
+//		} else if (blockState2.getFluidState().getAmount() == 8) {
+//			System.out.println("false at"+blockPos+" due to fluid");
+//			return false;
+//		} else if (blockState2.hasProperty(FallingLayer.LAYERS) && blockState2.getValue(FallingLayer.LAYERS) > 1) {
+//			System.out.println("false at"+blockPos+" due to layers and exceeding");
+//			return false;
+//		} else {
+//			int i = LightEngine.getLightBlockInto(levelReader, blockState, blockPos, blockState2, blockPos2, Direction.UP, blockState2.getLightBlock(levelReader, blockPos2));
+//			System.out.println((i < levelReader.getMaxLightLevel())+" at"+blockPos+" due to light ("+i+","+levelReader.getMaxLightLevel()+")");
+//			return i < levelReader.getMaxLightLevel();
+//		}
+//	}
+
 	@Unique private static boolean tough_terrain$canBeGrass(BlockState blockState, LevelReader levelReader, BlockPos blockPos) {
 		BlockPos blockPos2 = blockPos.above();
 		BlockState blockState2 = levelReader.getBlockState(blockPos2);
@@ -40,7 +59,6 @@ public abstract class SpreadingSnowyDirtBlockMixin extends SnowyDirtBlock {
 			return i < levelReader.getMaxLightLevel();
 		}
 	}
-
 	@Shadow private static boolean canPropagate(BlockState blockState, LevelReader levelReader, BlockPos blockPos) {
 		BlockPos blockPos2 = blockPos.above();
 		return canBeGrass(blockState, levelReader, blockPos) && !levelReader.getFluidState(blockPos2).is(FluidTags.WATER);
@@ -55,7 +73,7 @@ public abstract class SpreadingSnowyDirtBlockMixin extends SnowyDirtBlock {
 	private void randomTick(BlockState state, ServerLevel level, BlockPos pos, RandomSource random, CallbackInfo ci) {
 		if (!tough_terrain$canBeGrass(state, level, pos)) {
 			if (state.hasProperty(FallingLayer.LAYERS)) {
-				level.setBlockAndUpdate(pos, ModBlocks.LOOSE_DIRT.getPlacedLayer().defaultBlockState().setValue(FallingLayer.LAYERS,state.getValue(FallingLayer.LAYERS).intValue()));
+				level.setBlockAndUpdate(pos, ModBlocks.LOOSE_DIRT.getPlacedLayer().defaultBlockState().setValue(FallingLayer.LAYERS, state.getValue(FallingLayer.LAYERS)));
 				ci.cancel();
 			} else {
 				level.setBlockAndUpdate(pos, Blocks.DIRT.defaultBlockState());
